fix: reorganize call and client state
