+++ packages/client/src/store/index.ts
@@ -1,3 +1,4 @@
 export * as RxUtils from './rxUtils';
+export * from './CallingState';
 export * from './stateStore';
 export * from './CallState';
