+++ packages/react-bindings/src/hooks/store.ts
@@ -11,7 +11,7 @@ export const useStore = () => {
       `StreamVideoClient isn't initialized or this hook is called outside of <StreamVideo> context.`,
     );
   }
-  return client.readOnlyStateStore;
+  return client.state;
 };
 
 /**
